Move `RouterSpy` to private scope, since it's not shared anymore.
